backend/requirements.txt
fastapi==0.136.1
starlette==0.49.3
uvicorn[standard]==0.30.6
sqlalchemy==2.0.49
psycopg[binary]==3.3.3
pydantic-settings==2.6.0
email-validator==2.2.0
python-jose[cryptography]==3.5.0
passlib==1.7.4
python-multipart==0.0.26
pandas==2.3.3
openpyxl==3.1.5
requests==2.33.0
httpx==0.28.1
slowapi==0.1.9
structlog==24.4.0
sentry-sdk[fastapi]==2.18.0
alembic==1.13.3
pytest==9.0.3
pytest-cov==5.0.0
ruff==0.6.9
pylint==3.3.1
bandit==1.8.3
pip-audit==2.8.0
pre-commit==4.0.1
mypy==1.13.0
types-passlib==1.7.7.20240819
types-python-jose==3.3.4.20240106
types-requests==2.32.0.20241016
pyasn1==0.6.3
